service/requirements.txt
pixel-multiverse
PyYAML